Add files via upload
This is a project developed for arduino UNO with which we seek to automate the growth of a cannabis plant by controlling the number of hours of light, ambient temperature and humidity of the ground. displaying the data of the environment on an LCD screen at all times. connect the fan and water pump by means of a relay with 5v coil, the connection to the LCD screen is made by the I2C protocol.